Small Haskell Programs
Added in the handful of small Haskell programs I wrote over the course of
learning a bit of Haskell. I more or less just use GHCi as a functional
calculator, so these are all plain programs that don't make the slightest
use of external libraries.

Also, an Ada program forgotten in the last commit, which calculates
triangular-square numbers (named `triangle.adb`).

Added programs:

`hello.hs`: Obligatory "Hello, world!" program.

`sumorproduct.hs`: A little bit that reads a command and performs one of
two actions; a sum or a product over a number range starting at 1.

`guess.hs`: Another deterministic guessing game, this time in Haskell.

`triangle.hs`: Another triangular-square number calculator.

`sieve.hs`: A prime-number generator which uses a moderately optimized
Sieve of Eratosthenes to calculate primes.

`multiplicationtable.hs`: Since generating multiplication tables is so
deathly important in this day and age...

`rpnCalc.hs`: A reverse-polish-notation calculator, made using a stack of
items.

Scripts (for running w/the haskell interpreter):

`agm.hs`: A little script made implementing the arithmetic-geometric mean,
and an algorithm for calculating logarithms.

`euler.hs`: An implementation of Euler's method, as well as a slight
improvement to the method. I wrote this for differential equations
homework, since I didn't have Maple (the CAS) available at home. It's got a
clunky interface, but it works.

`lsystem.hs`: A little script for playing around with L-systems.

`symbolicMath.lhs`: This is pretty interesting, in that it's a first-level
attempt at writing the very back of a symbolic math engine from scratch. It
is *really* bare, but kind of neat for something I wrote with no
pre-existing knowledge on the subject.