Small Rust programs
A few tiny Rust programs. Note that I disobey the Rust formatting rules a
bit, in that I use full-width tabs.

Also, a forgotten Haskell script, `sumpair.hs`, which finds all pairs of a
given sum in a given array.

New Rust programs:

`triangles`: An implementation of triangular-square numbers in Rust --
basically the first thing I wrote in Rust, so it's probably got a few
things wrong with it.

`sumpair`: A rewrite of the `sumpair.hs`, but as a full working Rust
program.

Rust libraries:

`complex`: An implementation of complex-numbers. Not a very comprehensive
one, but a working one nontheless.

`fifo.rs`: A freestanding source file that implements a circular buffer
type -- that is, a buffer in which the newest elements pushed overwrite the
oldest, to work as a cache.